rearrange cli code
