+++ src/utils/cli/utils.py
@@ -340,3 +340,62 @@ def validate_args(
     args.output_directory.mkdir(parents=True, exist_ok=True)
 
     return args
+
+
+def route_pipeline_step(
+    args: argparse.Namespace,
+) -> int:
+    """Route pipeline step to either local execution or SLURM submission.
+
+    Args:
+        args: Arguments to validate. Must contain a pipeline_step_func attribute
+            that should handle saving its outputs.
+
+    Returns:
+        int: 0 if successful
+    """
+    args = validate_args(
+        args,
+        input_directory_name=args.input_directory_name,
+        output_directory_name=args.output_directory_name,
+    )
+
+    if args.slurm:
+        import submitit
+
+        print("Submitting to cluster compute nodes with submitit")
+        executor = submitit.AutoExecutor(folder="logs")
+        executor.update_parameters(
+            slurm_time=600,
+            slurm_cpus_per_task=1,
+            slurm_mem_per_cpu=256000,
+            slurm_array_parallelism=3,
+            slurm_partition="general",
+        )
+        with executor.batch():
+            for state in args.states:
+                args.state = state
+                executor.submit(args.pipeline_step_func, args)
+    else:
+        for state in args.states:
+            print(f"Running pipeline step for {state}")
+            args.state = state
+            args.pipeline_step_func(args)
+    return 0
+
+
+def create_subparsers(
+    parser: argparse.ArgumentParser,
+) -> dict[str, argparse.ArgumentParser]:
+    """Create subparsers for all pipeline steps.
+
+    Args:
+        parser: ArgumentParser to add subparsers to
+    Returns:
+        dict[str, argparse.ArgumentParser]: Dictionary of step name to parser
+    """
+    subparsers = parser.add_subparsers(dest="command", required=True)
+    step_parsers = {}
+    for step in pipeline_step_details:
+        step_parsers[step] = create_parser_for_step(subparsers[step], step)
+    return step_parsers
